+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod api;
-pub mod apis;
 pub mod cbor;
 pub mod config;
 pub mod dkg;
@@ -12,3 +11,4 @@ pub mod price_feed;
 pub mod publisher;
 pub mod raft;
 pub mod signature_aggregator;
+pub mod sources;
