+++ src/sources/maestro.rs
@@ -6,7 +6,7 @@ use std::{env, sync::Arc, time::Duration};
 use tokio::{task::JoinSet, time::sleep};
 use tracing::{warn, Instrument};
 
-use crate::apis::source::{PriceInfo, PriceSink};
+use crate::sources::source::{PriceInfo, PriceSink};
 
 use super::source::Source;
